refact: rename `upgrade_from_local_wheel` test suite
* Unit tests upgrading a python package from 2.0.0 -> 2.0.1
* Specify which dependencies are expected to exist in venv. For example
  `test_upgrade_local_wheel_top_level_package_2_0_0_where_package_name_is_valid_expect_package_installed`
  expects that oll-* packages are on 2.0.1 versions
